Improves organization of endpoints in controllers classes
- All methods with endpoints starting with "/topics"
  are now located in TopicRestController.
  Endpoint "/topics/{topicId}/polls" has been moved
  from PollRestController.

- All methods with endpoints starting with "/polls"
  are now located in PollRestController.
  Endpoint "/polls/{pollId}/votes" has been moved
  from VoteRestController.

- VoteRestController became empty, so it was deleted.
