+++ src/main/java/com/vollino/poll/service/poll/vote/VoteRepository.java
@@ -1,4 +1,4 @@
-package com.vollino.poll.service.vote;
+package com.vollino.poll.service.poll.vote;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
