+++ src/test/java/com/vollino/poll/service/topic/TopicRestControllerTest.java
@@ -22,6 +22,7 @@ import java.util.List;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -62,7 +63,37 @@ public class TopicRestControllerTest {
     }
 
     @Test
-    public void shouldAllTopics() throws Exception {
+    public void shouldCreatePoll() throws Exception {
+        //given
+        Long topicId = 2L;
+        Poll received = new Poll(null, topicId, "Poll description",
+                ZonedDateTime.parse("2019-04-16T18:19:00-03:00[Brazil/East]"));
+        Poll persisted = new Poll(1L, topicId, "Poll description",
+                ZonedDateTime.parse("2019-04-16T18:19:00-03:00[Brazil/East]"));
+        given(pollService.create(received)).willReturn(persisted);
+
+        //when
+        ResultActions response = mockMvc.perform(post("/topics/{topicId}/polls", topicId)
+                .content("{" +
+                        "\"description\": \"Poll description\"," +
+                        "\"endDate\": \"2019-04-16T18:19:00-03:00[Brazil/East]\"" +
+                    "}")
+                .contentType(MediaType.APPLICATION_JSON_UTF8));
+
+        //then
+        response.andExpect(status().isCreated())
+                .andExpect(content().contentType(MediaType.APPLICATION_JSON_UTF8))
+                .andExpect(content().json("{" +
+                        "\"id\": 1," +
+                        "\"topicId\": 2," +
+                        "\"description\": \"Poll description\"," +
+                        "\"endDate\": \"2019-04-16T18:19:00-03:00[Brazil/East]\"" +
+                    "}"));
+        verify(pollService).create(received);
+    }
+
+    @Test
+    public void shouldGetAllTopics() throws Exception {
         //given
         Topic topic1 = new Topic(1L, "Topic 1 description");
         Topic topic2 = new Topic(2L, "Topic 2 description");
